Adds an example of modifications to an item provider to
a) change the textual representation of a class (modifying getText(...)
b) change the choice of values for properties (modifying
getChoiceOfValues(...) in the descriptor.

Also adds a group with a team to the example model, which allows to see
the difference in available choices for home or guest team in the editor
when the group of a match is not set.